Multi-display support. Works with different DPI per monitor
Also contains a few other drive-by changes and improvements as well as
some preparations for further changes.

Note that this changes requires that Windows knows that Kolonnade is
per-monitor DPI aware. This is done via the manifest, but when run out
of Rider, Windows thinks that Kolonnade is only system-wide DPI aware,
not per monitor. When run from the commandline (`dotnet run`) or out of
Visual Studio, everything works as it should.