Replace hotkey handling: assume another component sends messages
I really want the Win key as modifier, but unfortunately Windows doesn't
allow hotkeys with the Win key as modifier (something something
plattform only OEMs are allowed to use it). The alt key is used by many
regular app shortcuts and control + alt feels unhandy. Hence assume that
there is another component that notices a hotkey and sends a message
"KolonnadeHotKey" with the key as wparam and a possible mod mask as
lparam. In my case, it's my Neo2 low-level keyboard hook that does it.