Rework hotkey handling to work around SetForegroundWindow restrictions
Windows doesn't allow every app to set arbitrary windows to foreground.
Instead, there are certain conditions where an app is allowed to bring
a window to foreground. One situation where Windows allows an app to set
the foreground window if it reacts to a hotkey. On the other hand,
Windows doesn't allow hotkeys that involve the win key. Therefore,
hotkey handling is two-fold: a low-level keyboard hook (outside of
Kolonnade) reacts on win key presses and creates a broadcast message with
the key. Kolonnade listens to the broadcast message and stores them in a
queue. The external component then injects key presses so Windows
activates Kolonnade as a hotkey. That way, Kolonnade has the permission
to use SetForegroundWindow and Kolonnade now processes the enqueued
hotkeys.